Fix AppVeyor build errors

+++ src/gui/src/KeySequence.cpp
@@ -233,7 +233,8 @@ QString KeySequence::keyToString(int key)
 
     // representable in ucs2?
     if (key < 0x10000)
-        return QString("\\u%1").arg(QChar(key).toLower().unicode(), 4, 16, QChar('0'));
+        return QString("\\u%1").arg(
+            static_cast<unsigned int>(QChar(key).toLower().unicode()), 4, 16, QChar('0'));
 
     // give up, InputLeap probably won't handle this
     return "";
